Rework structure to remove nesting for int/ext stg/prod folder
Go back to the original design to have both ext/int and stg/prd
information as part of the overlay. This is needed to deploy components
on all cluster that require specific whether they are on internal or
external cluster.

Align both argo-cd-apps overlays and components folders to follow that
pattern. To achieve this in argo-cd-apps, I needed to move the
definition of the applicationsets into a base folder and group them in
external, internal or all-clusters folders based on where we need them.
This is same pattern as we do in original infra-deployments.

Signed-off-by: Hugo Arès <hares@redhat.com>