Set ApplicationSet default environment to empty string
The environment is getting changed by a patch to the right value, i.e.

* internal-staging
* internal-production
* external-staging
* external-production

Change the ApplicationSets to have nothing as default to remove
confusion but also in case we do not patch it, it will not deploy thing
from unexpected folder as the root folder of components usually does not
contains kustomization.yaml.

Signed-off-by: Hugo Arès <hares@redhat.com>