ci: add yamllint to checks
Add a check to our CI to ensure that our manifests comply with our
yamllint config.

Similar to infra-deployments, we should ensure that our manifests are
valid yaml documents and that our manifests have some amount of
formatting consistency.  For portability purposes, use the same yamllint
configuration that infra-deployments uses.

Part-of: KFLUXINFRA-2284
Signed-off-by: Andy Sadler <ansadler@redhat.com>