chore(RELEASE-2005): swap standard internal-services radas stage cert
This commit updates the internal-services ExternalSecrets for both prod
and stage to change the key used for the hacbs signing pipeline stage
certs. If all goes well with this merged, the prod one will be changed
as well.

The point of this change is to not maintain the same secret in two
separate vault locations.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>